Rename gameProcessors to game-processor-utils and refactor imports
